+++ scripts/RigidBodyHitTest.cs
@@ -2,9 +2,9 @@
 
 public class RigidBodyHitTest : RigidBody
 {
-    private const int BASE_BULLET_BOOST = 9;
+    private const byte BASE_BULLET_BOOST = 9;
 
-    public void BulletHit(int damage, Transform bulletGlobalTransform)
+    public void BulletHit(byte damage, Transform bulletGlobalTransform)
     {
         var directionVector = bulletGlobalTransform.basis.z.Normalized() * BASE_BULLET_BOOST;
         ApplyImpulse(
